builders: Use netifaces to get docker host IP
This will allow better portability on various OS (MacOS for example).
the `ip` command is only available on GNU/Linux and the `ifconfig` is
available on all Unix system but is deprecate on GNU/Linux.

Netifaces work on GNU/Linux, OS X and Microsoft Windows. It should also
work on other Unix system but they are not officialy supported.