builders: Fix docker-machine support
When using Docker Machine, the docker server certificate commonName or
subjectAltName (it's boot2docker) do not match the docker server
hostname (it's server ip), so we have to disable hostname validation to
allow docker server usage.

Also, Docker Machine add an intermediate VM between the docker server
and the host machine. In this VM only the home directory (in fact, the
parent directory of the home directory) of the current user is mounted.
So we can not use volume mount from files outside of the home directory.