six: Silently fail when using sync whithout libc 6
In Python 3, the standard library allow you to force flush data to disk
using the os.sync() command. But this function does not exist on Python 2.
We have a compatibility function six.sync() but it use libc.so.6 and
so it is not compatible with OS X. On OS X, the "text file busy" issue
has never append, so we do not have to find a sync replacement for this
OS.