Compiler: Stop mounting pip config file
When using docker volume to expose host pip config file in container,
the file keep its owner and acls. When your user has the same id as
grocker user in the container (ie 1000), it's not a problem but in the
other case you are unable to build a Grocker image for private packages.

This commit change the described beaviour to use env vars to set pip
config on the compiler container.

refs #84