Retrive image manifest hash
Since at least Docker registry 2.6.0 the image digest and the manifest
digest differ. The Grocker result file give the digest to allow checking
it during the pull process. With the new registry this always fail.

For private registry, we now retrieve the manifest digest using the http
API to fix this issue.