builders: Use extra requirement when building from a wheel filepath
Commit 7dd124c0c9dc7c47ff57e722228b4c4373a3e6a3 introduced a way to pass
a wheel path as a release argument. However, it ignored the extra
requirements that could be provided with it, e.g.:

    /path/to/grocker_test_project-3.0.1-py2.py3-none-any.whl[pep8]

This commit introduces a new GrockerRequirement property `pip_extras`
which returns the extras ready to be passed to pip (e.g.: '[pep8]'). The
property is used to fix grocker.builders.compile_wheels.

Update doc and CLI help with information about extra requirements.